Fix -Wsign-compare warnings throughout the codebase
Resolve all -Wsign-compare warnings found when building with
-Wall -Wsign-compare -Wno-unused-parameter -DCMAKE_UNITY_BUILD=OFF.

Strategy used per-site:
- Changed loop counters from GLint/int to GLuint where the variable
  only iterates over unsigned counts (vbo_exec_array.c, api_validate.c,
  context.c, slang_print.c, slang_vartable.c, ss_context.c,
  prog_execute.c).
- Changed GLint min_index/max_index to GLuint in vbo_split_inplace.c
  since they are only assigned from unsigned struct fields.
- Added (GLuint) or (GLint) casts at comparison sites where changing
  the variable type is unsafe or touches public API (shader_api.c,
  slang_codegen.c, slang_compile.c, slang_emit.c, slang_link.c,
  slang_simplify.c, slang_vartable.c, prog_execute.c, texstore.c,
  renderbuffer.c, arbprogparse.c, t_draw.c).
- Cast sizeof/array-size expressions to (GLint) in assert() calls
  in prog_print.c.
- Fixed ternary operator signedness mismatch in slang_emit.c.
- Changed ~0 to ~0u for unsigned dispatch_offset comparisons in
  glapi.c.
- Added (GLint) casts for gl_state_index enum vs int comparisons in
  arbprogparse.c.

No struct field types or function signatures were changed.

+++ src/main/renderbuffer.c
@@ -2122,9 +2122,9 @@ _mesa_reference_renderbuffer(struct gl_renderbuffer **ptr,
 	GLboolean deleteFlag = GL_FALSE;
 	struct gl_renderbuffer *oldRb = *ptr;
 
-	assert(oldRb->Magic == RB_MAGIC);
+	assert((unsigned int)oldRb->Magic == RB_MAGIC);
 	_glthread_LOCK_MUTEX(oldRb->Mutex);
-	assert(oldRb->Magic == RB_MAGIC);
+	assert((unsigned int)oldRb->Magic == RB_MAGIC);
 	ASSERT(oldRb->RefCount > 0);
 	oldRb->RefCount--;
 	/*printf("RB DECR %p (%d) to %d\n", (void*) oldRb, oldRb->Name, oldRb->RefCount);*/
@@ -2141,7 +2141,7 @@ _mesa_reference_renderbuffer(struct gl_renderbuffer **ptr,
     assert(!*ptr);
 
     if (rb) {
-	assert(rb->Magic == RB_MAGIC);
+	assert((unsigned int)rb->Magic == RB_MAGIC);
 	/* reference new renderbuffer */
 	_glthread_LOCK_MUTEX(rb->Mutex);
 	rb->RefCount++;
