+++ src/shader/slang/slang_vartable.c
@@ -98,12 +98,12 @@ void
 _slang_pop_var_table(slang_var_table *vt)
 {
     struct table *t = vt->Top;
-    int i;
+    GLuint i;
 
     if (dbg) printf("Popping level %d\n", t->Level);
 
     /* free the storage allocated for each variable */
-    for (i = 0; i < t->NumVars; i++) {
+    for (i = 0; i < (GLuint)t->NumVars; i++) {
 	slang_ir_storage *store = (slang_ir_storage *) t->Vars[i]->aux;
 	GLint j;
 	GLuint comp;
@@ -204,18 +204,18 @@ alloc_reg(slang_var_table *vt, GLint size, GLboolean isTemp)
 
     for (i = 0; i <= vt->MaxRegisters * 4 - size; i += step) {
 	GLuint found = 0;
-	for (j = 0; j < size; j++) {
+	for (j = 0; j < (GLuint)size; j++) {
 	    if (i + j < vt->MaxRegisters * 4 && t->Temps[i + j] == FREE) {
 		found++;
 	    } else {
 		break;
 	    }
 	}
-	if (found == size) {
+	if (found == (GLuint)size) {
 	    /* found block of size free regs */
 	    if (size > 1)
 		assert(i % 4 == 0);
-	    for (j = 0; j < size; j++)
+	    for (j = 0; j < (GLuint)size; j++)
 		t->Temps[i + j] = isTemp ? TEMP : VAR;
 	    t->ValSize[i] = size;
 	    return i;
@@ -300,7 +300,7 @@ _slang_free_temp(slang_var_table *vt, slang_ir_storage *store)
     } else {
 	/*assert(store->Swizzle == SWIZZLE_NOOP);*/
 	assert(t->ValSize[r*4] == store->Size);
-	for (i = 0; i < store->Size; i++) {
+	for (i = 0; i < (GLuint)store->Size; i++) {
 	    assert(t->Temps[r * 4 + i] == TEMP);
 	    t->Temps[r * 4 + i] = FREE;
 	}
@@ -314,7 +314,7 @@ _slang_is_temp(const slang_var_table *vt, const slang_ir_storage *store)
     struct table *t = vt->Top;
     GLuint comp;
     assert(store->Index >= 0);
-    assert(store->Index < vt->MaxRegisters);
+    assert((GLuint)store->Index < vt->MaxRegisters);
     if (store->Swizzle == SWIZZLE_NOOP)
 	comp = 0;
     else
